Reworked task 3.
first part of task 3 (the 1998 top 300 ranking) was indeed missing, i added it in. but i also reworked the last SELECT of this query, because i only now realised that the "amount_sold" column should have the cross-year totals for each channel & customer, not the total for one of the years they ranked top 300. (again, the instructions on learn.epam were very ambigous)
P.S. Today i got an email inviting me to stage 2 without the pre-lab interview, so i want to thank you for your kind reviews! hope to meet you there.